+++ irodsbackend/templatetags/irodsbackend_tags.py
@@ -1,17 +1,17 @@
 from django import template
-from django.conf import settings
 from django.urls import reverse
 from django.utils.http import urlencode
 
-from ..api import IrodsAPI
+from irodsbackend.api import IrodsAPI
+
+# Samplesheets dependency
+from samplesheets.utils import get_webdav_url as _get_webdav_url
 
 
 irods_backend = IrodsAPI(conn=False)
-
 register = template.Library()
 
 
-# TODO: Not needed anymore, can be removed
 @register.simple_tag
 def get_irods_path(obj):
     return irods_backend.get_path(obj)
@@ -20,7 +20,8 @@ def get_irods_path(obj):
 @register.simple_tag
 def get_stats_html(irods_path, project):
     """
-    Return collection stats badge element into a template
+    Return collection stats badge element into a template.
+
     :param irods_path: Full iRODS path (string)
     :param project: Project object
     :return: String (contains HTML)
@@ -33,7 +34,7 @@ def get_stats_html(irods_path, project):
         + urlencode(query_string)
     )
     return (
-        '<span class="badge badge-pill badge-info sodar-irods-stats"'
+        '<span class="badge badge-pill badge-info sodar-irods-stats" '
         'data-stats-url="{}">'
         '<i class="iconify spin" data-icon="mdi:loading"></i> Updating..'
         '</span>'.format(url)
@@ -41,23 +42,8 @@ def get_stats_html(irods_path, project):
 
 
 @register.simple_tag
-def is_webdav_enabled():
-    return settings.IRODS_WEBDAV_ENABLED
-
-
-@register.simple_tag
-def get_webdav_url():
-    if settings.IRODS_WEBDAV_ENABLED:
-        return settings.IRODS_WEBDAV_URL.rstrip('/')
-
-
-@register.simple_tag
-def get_webdav_url_anon():
-    if settings.IRODS_WEBDAV_ENABLED:
-        return settings.IRODS_WEBDAV_URL_ANON.rstrip('/')
-
-
-@register.simple_tag
-def get_webdav_user_anon():
-    if settings.IRODS_WEBDAV_ENABLED:
-        return settings.IRODS_WEBDAV_USER_ANON
+def get_webdav_url(project, user):
+    url = _get_webdav_url(project, user)
+    if not url:
+        return ''
+    return url